+++ client/src/components/common/Loader.js
@@ -4,10 +4,10 @@ import Box from "@mui/material/Box";
 
 const Loader = () => {
   return (
-    <Box
-      sx={{ display: "flex" }}
-      style={{ flexDirection: "column", alignItems: "center" }}
-    >
+    <Box 
+    display="flex"
+    flexDirection="column" 
+    alignItems="center">
       <CircularProgress />
       <p>Loading...</p>
     </Box>
