--- client/src/components/common/Loader/index.js
@@ -1,3 +1,0 @@
-import Loader from "./Loader";
-
-export default Loader;
