--- client/src/components/common/ProductsFilter/PopUpFilter/index.js
@@ -1,3 +1,0 @@
-import PopUpFilter from "./PopUpFilter";
-
-export default PopUpFilter;
