+++ client/src/redux/actions/cartActions.js
@@ -1,5 +1,5 @@
 import ProductService from '../../services/ProductService';
-import { CART_ADD_ITEM, CART_REMOVE_ITEM } from '../constants/cartConstants';
+import { CART_ADD_ITEM, CART_REMOVE_ITEM, CART_ADD_ITEM_QTY, CART_REMOVE_ITEM_QTY} from '../constants/cartConstants';
 
 export const addToCart = (id, qty = 1) => async (dispatch, getState) => {
   const data = await ProductService.getById(id);
@@ -11,6 +11,8 @@ export const addToCart = (id, qty = 1) => async (dispatch, getState) => {
       name: data.name,
       image: data.imageUrls[0],
       price: data.currentPrice,
+      brand: data.brand,
+      color: data.color,
       countInStock: data.quantity,
       qty,
     },
@@ -28,5 +30,20 @@ export const removeFromCart = (id) => (dispatch, getState) => {
     payload: id,
   });
 
+  localStorage.setItem('cartItems', JSON.stringify(getState().cart.cartItems));
+};
+export const removeFromCartItemQty = (id) => (dispatch, getState) => {
+  dispatch({
+    type: CART_REMOVE_ITEM_QTY,
+    payload: id,
+  });
+
+  localStorage.setItem('cartItems', JSON.stringify(getState().cart.cartItems));
+};
+export const addToCartItemQty = (id) => (dispatch, getState) => {
+  dispatch({
+    type: CART_ADD_ITEM_QTY,
+    payload: id,
+  });
   localStorage.setItem('cartItems', JSON.stringify(getState().cart.cartItems));
 };
